shared: Moves ExitStatus into unix build file
