+++ shared/util_unix.go
@@ -3,8 +3,12 @@
 package shared
 
 import (
+	"errors"
 	"os"
+	"os/exec"
 	"syscall"
+
+	"golang.org/x/sys/unix"
 )
 
 func GetOwnerMode(fInfo os.FileInfo) (os.FileMode, int, int) {
@@ -12,4 +16,30 @@ func GetOwnerMode(fInfo os.FileInfo) (os.FileMode, int, int) {
 	uid := int(fInfo.Sys().(*syscall.Stat_t).Uid)
 	gid := int(fInfo.Sys().(*syscall.Stat_t).Gid)
 	return mode, uid, gid
+}
+
+// ExitStatus extracts the exit status from the error returned by exec.Cmd.
+// If a nil err is provided then an exit status of 0 is returned along with the nil error.
+// If a valid exit status can be extracted from err then it is returned along with a nil error.
+// If no valid exit status can be extracted then a -1 exit status is returned along with the err provided.
+func ExitStatus(err error) (int, error) {
+	if err == nil {
+		return 0, err // No error exit status.
+	}
+
+	var exitErr *exec.ExitError
+
+	// Detect and extract ExitError to check the embedded exit status.
+	if errors.As(err, &exitErr) {
+		// If the process was signaled, extract the signal.
+		status, isWaitStatus := exitErr.Sys().(unix.WaitStatus)
+		if isWaitStatus && status.Signaled() {
+			return 128 + int(status.Signal()), nil // 128 + n == Fatal error signal "n"
+		}
+
+		// Otherwise capture the exit status from the command.
+		return exitErr.ExitCode(), nil
+	}
+
+	return -1, err // Not able to extract an exit status.
 }
